modify 100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -11,5 +11,5 @@
 
 int _atoi(char *s)
 {
-	return (atoi(s));
+	return (atol(s));
 }
